@@ -1,4 +1,5 @@
 import { GraphQLNonNull, GraphQLEnumType } from 'graphql';
+import deepcopy from 'deepcopy';
 import { mutationWithClientMutationId } from 'graphql-relay';
 import { FunctionsRouter } from '../../Routers/FunctionsRouter';
 import * as defaultGraphQLTypes from './defaultGraphQLTypes';
@@ -43,7 +44,7 @@ const load = parseGraphQLSchema => {
       },
       mutateAndGetPayload: async (args, context) => {
         try {
-          const { functionName, params } = args;
+          const { functionName, params } = deepcopy(args);
           const { config, auth, info } = context;
 
           return {
